added variables, google fonts, css for buttons

+++ react/redux-posts/app/shared/DateDisplayer/DateDisplayer.tsx
@@ -26,7 +26,7 @@ function DateDisplayer({date} : {date: string}) {
   } else {
     renderedDate = dateOfEvent.toLocaleString(
       undefined,
-      {year: "numeric", month: "long", day: "2-digit", hour: "2-digit", minute: "2-digit"}
+      {year: "numeric", month: "short", day: "2-digit", hour: "2-digit", minute: "2-digit"}
     );
   }
